backend/requirements.txt
alembic==1.12.1
bcrypt==4.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
filelock==3.12.2
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.0.5
future==1.0.0
gevent==22.10.2
greenlet==3.0.3
idna==3.7
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.4
Mako==1.2.4
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
PyJWT==2.8.0
PySocks==1.7.1
python-dotenv==0.21.1
requests==2.31.0
requests-file==2.1.0
SQLAlchemy==2.0.32
tldextract==4.0.0
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.15.0
zope.event==5.0
zope.interface==6.4.post2
